requirements.txt
aldryn-apphooks-config==0.3.3
aldryn-boilerplates==0.7.5
aldryn-common==1.0.4
aldryn-search==0.3.0
appdirs==1.4.3
Babel==2.3.4
beautifulsoup4==4.5.3
chardet==2.3.0
cmsplugin-filer==1.1.3
dj-database-url==0.4.2
Django==1.9.9
django-app-namespace-template-loader==0.4.1
django-appconf==1.0.2
django-appdata==0.1.6
django-blog-zinnia==0.18.1
django-bootstrap-themes==3.3.6
django-classy-tags==0.8.0
django-cms==3.4.2
django-compressor==2.1
django-contrib-comments==1.8.0
django-crispy-forms==1.6.1
django-document-library==1.5.5
django-extra-views==0.6.4
django-filer==1.2.7
django-filter==1.0.1
django-formtools==2.0
django-guardian==1.4.6
django-haystack==2.6.0
django-hvad==1.7.0
django-libs==1.67.12
django-localized-names==0.2.1
django-meta==1.3.2
django-meta-mixin==0.3.0
django-mptt==0.8.7
django-multilingual-tags==0.9.1
django-nose==1.4.4
django-oscar==1.4
django-oscar-api==1.1.0
django-parler==1.7
django-people==1.3
django-polymorphic==1.0.2
django-sekizai==0.10.0
django-sortedm2m==1.3.3
django-spurl==0.6.4
django-standard-form==1.1.1
django-tables2==1.4.2
django-tagging==0.4.5
django-taggit==0.22.0
django-taggit-autosuggest==0.3.0
django-taggit-helpers==0.1.4
django-taggit-templatetags==0.2.5
django-templatetag-sugar==1.0
django-treebeard==4.1.0
django-widget-tweaks==1.4.1
django-xmlrpc==0.1.7
djangocms-admin-style==1.2.6.2
djangocms-apphook-setup==0.1.2
djangocms-attributes-field==0.2.0
djangocms-blog==0.8.11
djangocms-column==1.7.0
djangocms-googlemap==1.0.2
djangocms-installer==0.9.5
djangocms-link==2.1.1
djangocms-snippet==1.9.2
djangocms-style==2.0.2
djangocms-text-ckeditor==3.4.0
djangocms-video==2.0.3
djangorestframework==3.6.2
dsdev-utils==0.9.6
easy-thumbnails==2.3
factory-boy==2.8.1
fake-factory==0.7.2
Faker==0.7.10
funcsigs==1.0.2
html5lib==0.9999999
ipaddress==1.0.18
lxml==3.7.3
Markdown==2.6.8
Menus==0.2.0
mock==2.0.0
mots-vides==2015.5.11
nose==1.3.7
olefile==0.44
packaging==16.8
pbr==2.0.0
phonenumbers==8.3.3
Pillow==4.0.0
purl==1.3
pycountry==1.20
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2016.10
rcssmin==1.0.6
regex==2017.2.8
rjsmin==1.0.12
six==1.10.0
sorl-thumbnail==12.4a1
treebeard==0.1.1
tzlocal==1.3
Unidecode==0.4.20
URLObject==2.4.2
YURL==0.13
